requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-healthcheck==1.7.0
six==1.11.0
Werkzeug==0.14.1
